Added interface/impl for service.
